Finalized algorithm color range
