+++ Files/sandbox.py
@@ -2,63 +2,45 @@ import cv2
 import numpy as np
 from scipy import ndimage
 times=0
-vs=cv2.VideoCapture("/Users/aidanobrien/Documents/GitHub/BeeFanningDetector/Assets/test_img&videos/13-21-55.h264")
-bk=cv2.imread('/Users/aidanobrien/Documents/GitHub/BeeFanningDetector/Assets/test_img&videos/testbkgrd1.jpg')
-bk2=cv2.imread('/Users/aidanobrien/Documents/GitHub/BeeFanningDetector/Assets/test_img&videos/black.png')
-fps = vs.get(cv2.CAP_PROP_FPS)
-print(fps)
+vs=cv2.VideoCapture("/Users/aidanobrien/Documents/GitHub.nosync/BeeFanningDetector/Assets/test_img&videos/90test1.h264")
+bk=cv2.imread('/Users/aidanobrien/Documents/GitHub.nosync/BeeFanningDetector/Assets/test_img&videos/90test1.png')
+bk2=cv2.imread('/Users/aidanobrien/Documents/GitHub.nosync/BeeFanningDetector/Assets/test_img&videos/black.png')
+#bk=bk[100:100+240,0:0+640]
+#bk2=bk2[100:100+240,0:0+640]
 while True:
     hasframes,img=vs.read()
-    
+    if(hasframes == False):
+        break
 
-    cv2.imshow("Output", img)
-    '''
+    
+    #img=img[100:100+240,0:0+640]
+    cv2.imshow("frame",img)
     subImage=(bk.astype('int32')-img.astype('int32')).clip(0).astype('uint8')
     grey=cv2.cvtColor(subImage,cv2.COLOR_BGR2GRAY)
     retval,thresh=cv2.threshold(grey,35,255,cv2.THRESH_BINARY)
-    #thresh=255-thresh
     kernel=np.ones((5,5),np.uint8)
     thresh=cv2.morphologyEx(thresh,cv2.MORPH_OPEN,kernel)
-    res2 = cv2.bitwise_and(img, img, mask= thresh)
-    cv2.imshow('Result2',res2)
+    noback = cv2.bitwise_and(img, img, mask= thresh)
+    #Color Bounds for 60 fps vids
+    #upper = np.array([220,220,220])  
+    #lower = np.array([160,160,160])  
 
-    upper = np.array([255,255,255])  #-- Lower range --
-    lower = np.array([128,128,128])  #-- Upper range --
-    mask = cv2.inRange(res2, lower, upper)
-    res = cv2.bitwise_and(res2, res2, mask= mask)  #-- Contains pixels having the gray color--
-    
+    #Color Bounds for 30 fps vids
+    upper = np.array([255,255,255])  
+    lower = np.array([128,128,128])  
+    mask = cv2.inRange(noback, lower, upper)
 
-    subImage2=(res.astype('int32')-bk2.astype('int32')).clip(0).astype('uint8')
-    grey2=cv2.cvtColor(subImage2,cv2.COLOR_BGR2GRAY)
-    retval2,thresh2=cv2.threshold(grey2,35,255,cv2.THRESH_BINARY)
-    #thresh2=255-thresh2
-    kernel2=np.ones((5,5),np.uint8)
-    thresh2=cv2.morphologyEx(thresh2,cv2.MORPH_OPEN,kernel)
-    im2, contours1, hierarchy1 = cv2.findContours(thresh2, cv2.RETR_TREE, cv2.CHAIN_APPROX_NONE)
-    cnt2=[]
-    for c in contours1:
-        
-        
-        x,y,w,h=cv2.boundingRect(c)
-        if(w*h>150 and w*h < 200 and w > h):
-            #cv2.rectangle(img,(x,y),(x+w,y+h),(0,255,0),2)
-            ell=cv2.fitEllipse(c)
-            cv2.ellipse(img,ell,(0,255,0),2)
-            print(w*h)
-        else:
-            cnt2.append(c)
-        
-    cv2.drawContours(thresh2, cnt2, -1, (0,0,0), cv2.FILLED)
-    cv2.imshow('Result',img)
-    cv2.imshow('Thresh',thresh2)
-    '''
+    wings = cv2.bitwise_and(noback, noback, mask=mask)
+    cv2.imshow('No back',noback)
+    cv2.imshow('Just_Wings/Shadows',wings)
     
     if times > 0:
-        key=cv2.waitKey(1) & 0xFF
+        key=cv2.waitKey(0) & 0xFF
         #if q is pressed, stop loop
         if key == ord("c"):
             continue
         if key == ord("q"):
             break
     times = times + 1
-    
+vs.release()
+cv2.destroyAllWindows()
